GCI: Figure out how checkpoints are stored for Spyro; there's a saved checkpoint for each map, it seems.
https://cdn.discordapp.com/attachments/696134617723437118/1056661827485970493/imagen.png
https://cdn.discordapp.com/attachments/696134617723437118/1056661984218718218/imagen.png

For Sphinx each save monkey statue has a unique number per level, so the save only needs to store that you are in the level with hashcode 0x100001 in the save point 3. Two uint32s. If they made it like this was for a reason.
The save system is completely different; much more complicated. They must have originally wanted to have arbitrary resume points between maps.